project/SlackToConfl/requirements.txt
slack-sdk>=3.27.0
anthropic>=0.40.0
atlassian-python-api>=3.41.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
